// chroma/src/chroma/types.ts
export type ExportTab = 'hex' | 'css' | 'array' | 'scss' | 'figma' | 'tailwind'